modeling/valueverse_engine/analysis_engine/requirements.txt
requests
psycopg2-binary
openai
python-dotenv
tenacity
